requirements/tests.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/tests.in
#
attrs==22.1.0
    # via pytest
coverage==6.0b1
    # via
    #   -r requirements/tests.in
    #   pytest-cov
exceptiongroup==1.0.4
    # via pytest
factory-boy==3.2.1
    # via pytest-factoryboy
faker==15.3.3
    # via factory-boy
flake8==4.0.1
    # via
    #   -r requirements/tests.in
    #   pytest-flake8
freezegun==1.2.2
    # via -r requirements/tests.in
idna==3.4
    # via yarl
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==1.1.1
    # via pytest
mccabe==0.6.1
    # via flake8
multidict==6.0.2
    # via yarl
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via
    #   -r requirements/tests.in
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
    #   pytest-flake8
    #   pytest-pythonpath
    #   pytest-vcr
pytest-cov==2.11.0
    # via -r requirements/tests.in
pytest-django==3.9.0
    # via -r requirements/tests.in
pytest-factoryboy==2.1.0
    # via -r requirements/tests.in
pytest-flake8==1.1.1
    # via -r requirements/tests.in
pytest-pythonpath==0.7.3
    # via -r requirements/tests.in
pytest-vcr==1.0.2
    # via -r requirements/tests.in
python-dateutil==2.8.2
    # via
    #   faker
    #   freezegun
pyyaml==6.0
    # via vcrpy
six==1.16.0
    # via
    #   python-dateutil
    #   vcrpy
tomli==2.0.1
    # via pytest
vcrpy==4.0.2
    # via
    #   -r requirements/tests.in
    #   pytest-vcr
wrapt==1.14.1
    # via vcrpy
yarl==1.8.1
    # via vcrpy
